CartManager - updateCart
